LayoutTests/fast/dom/NodeIterator/NodeIterator-dont-overcollect-expected.txt

This test asserts that NodeIterator's callback is kept alive while NodeIterator is alive even if all JS references go away.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS callbackWasTriggered is true
PASS successfullyParsed is true

TEST COMPLETE
